Update documentation, and bug correction
Started updating documentation, including installation of IDEA-CMWP in
Windows. Still incomplete.

Changed the name of the input files used as templates for IDEA. Added
numbers in their name so people know in wich order stuff should be
runned.

Changed the way that de IRF file is read, in order to solve (hopefully) a Windows
bug. It has been tried in Linux and everything works as always, but I
still need to test it in Windows.