Updates to IDEA.
Added the thickness correction for circular shape samples. Also added a
check in case the user made an error setting the sample.

Change the data range of the fitting. The first and last background
points define the range of the fitting. Its tested and it works.

Updated the IDEA documentation with this changes.